Specify a default value for Optional fields in models
The semantics of Optional were slightly altered in pydantic v2.
It seems earlier it treated it as a field that was optional and could
also be null. But in v2 it is a required field that can be null. To
maintain the schema we expect, we need to set a default value. The
validity of this change is confirmed by a couple of test cases which
would otherwise fail:

- test_content_load_failure_event_optional_error_field fails since it
  does not pass the optional field
- test_content_load_failed_v1_avro_schema fails since the avro schema
  will not generate "default": null for the field.